requirements_coolsys.txt
numpy==1.24.3
pyodbc==4.0.39
pandas==2.0.3
matplotlib==3.7.1
seaborn==0.12.2
scikit-learn==1.2.2
jupyter==1.0.0
notebook==7.0.6
ipykernel==6.25.0
PyPDF2==3.0.1
reportlab==4.0.4
torch>=2.0.0
transformers>=4.30.0
nltk>=3.8.1
tqdm>=4.65.0
pyspellchecker>=0.7.2
